feat(optimizer): adds support for function composition
