+++ compilers/concrete-optimizer/charts/src/bin/norm2_complexity.rs
@@ -47,6 +47,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         ciphertext_modulus_log,
         fft_precision,
         complexity_model: &CpuComplexity::default(),
+        composable: false,
     };
 
     let cache = decomposition::cache(
