+++ compilers/concrete-optimizer/concrete-optimizer-cpp/tests/src/main.cpp
@@ -29,6 +29,7 @@ concrete_optimizer::Options default_options() {
       .cache_on_disk = true,
       .ciphertext_modulus_log = CIPHERTEXT_MODULUS_LOG,
       .fft_precision = 53,
+      .composable = false
   };
 }
