+++ compilers/concrete-optimizer/v0-parameters/benches/benchmark.rs
@@ -20,6 +20,7 @@ fn v0_pbs_optimization(c: &mut Criterion) {
         cache_on_disk: true,
         ciphertext_modulus_log: 64,
         fft_precision: 53,
+        composable: false,
     };
 
     c.bench_function("v0 PBS table generation", |b| {
@@ -46,6 +47,7 @@ fn v0_pbs_optimization_simulate_graph(c: &mut Criterion) {
         cache_on_disk: true,
         ciphertext_modulus_log: 64,
         fft_precision: 53,
+        composable: false,
     };
 
     c.bench_function("v0 PBS simulate dag table generation", |b| {
@@ -72,6 +74,7 @@ fn v0_wop_pbs_optimization(c: &mut Criterion) {
         cache_on_disk: true,
         ciphertext_modulus_log: 64,
         fft_precision: 53,
+        composable: false,
     };
 
     c.bench_function("v0 WoP-PBS table generation", |b| {
